Separated the first mentionable command for things like headpat. Had to redo how jsoup found and parsed information from tenor because tenor's website changed and it wasn't properly grabbing the links anymore for the sites. Don't know why it freaked out so bad but it's fixed now thank fuck
