+++ bin/Resource/mentionUserList.txt
@@ -1,13 +1,1 @@
-pat
-boob
-slap
 hug
-snuggle
-kiss
-poke
-stare
-smug
-lewd
-bite
-meow
-butt
