+++ bin/Storage/Users/uwuData/285636067535880192.json
@@ -1,1 +1,1 @@
-{"Fine Amount":19950}
+{"Fine Amount":22400}
