+++ bin/Storage/Users/uwuData/714907896076435476.json
@@ -1,1 +1,1 @@
-{"Fine Amount":768950}
+{"Fine Amount":770700}
