+++ bin/main/.gitignore
@@ -1,2 +1,8 @@
 /Gaiza/
 /Listener/
+/Command/
+/Management/
+/UserCommands/
+/AdminCommands/
+/Database/
+/UserMentions/
